# swagger.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost:8080/api",
    "description" : "Generated server url"
  } ],
...
  "paths" : {
    "/user/update" : {
      "put" : {
        "tags" : [ "用户管理" ],
        "summary" : "更新用户信息",
        "description" : "更新用户个人资料信息",
        "operationId" : "updateUser",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserInfoDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
...
    "/waitlist/paid/{waitlistNumber}" : {
      "post" : {
        "tags" : [ "候补购票" ],
        "summary" : "候补订单支付成功回调",
        "description" : "候补订单支付成功后，系统处理出票",
        "operationId" : "waitlistOrderPaid",
        "parameters" : [ {
          "name" : "waitlistNumber",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultBoolean"
                }
              }
            }
          }
        }
      }
    },
    "/waitlist/create" : {
      "post" : {
        "tags" : [ "候补购票" ],
        "summary" : "创建候补购票请求",
        "description" : "用户提交候补购票请求，指定车次、座位类型、出行日期",
        "operationId" : "createWaitlistOrder",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/WaitlistCreateDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultWaitlistOrderDTO"
                }
              }
            }
          }
        }
      }
    },
    "/waitlist/cancel/{id}" : {
      "post" : {
        "tags" : [ "候补购票" ],
        "summary" : "取消候补订单",
        "description" : "取消用户提交的候补购票请求",
        "operationId" : "cancelWaitlistOrder",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultBoolean"
                }
              }
            }
          }
        }
      }
    },
    "/user/register" : {
      "post" : {
        "tags" : [ "用户管理" ],
        "summary" : "用户注册",
        "description" : "用户提供用户名、密码、手机号、身份证进行注册",
        "operationId" : "register",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserRegisterDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultUserInfoDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/user/login" : {
      "post" : {
        "tags" : [ "用户管理" ],
        "summary" : "用户登录",
        "description" : "用户名密码登录，返回JWT Token",
        "operationId" : "login",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserLoginDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
...
    "/waitlist/{id}" : {
      "get" : {
        "tags" : [ "候补购票" ],
        "summary" : "根据ID查询候补订单详情",
        "description" : "通过候补订单ID获取详细信息",
        "operationId" : "getWaitlistOrderById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultWaitlistOrderDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/waitlist/query" : {
      "get" : {
        "tags" : [ "候补购票" ],
        "summary" : "查询候补池可用票",
        "description" : "按条件筛选候补池中当前可用的车票",
        "operationId" : "queryAvailableTickets",
        "parameters" : [ {
          "name" : "trainId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "startStationId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "endStationId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "seatType",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "departureDate",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultListWaitlistTicketDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/user/check/username" : {
      "get" : {
        "tags" : [ "用户管理" ],
        "summary" : "检查用户名是否存在",
        "description" : "验证用户名是否已被注册",
        "operationId" : "checkUsername",
        "parameters" : [ {
          "name" : "username",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultBoolean"
                }
              }
            }
          }
        }
      }
    },
    "/user/check/phone" : {
      "get" : {
        "tags" : [ "用户管理" ],
        "summary" : "检查手机号是否存在",
        "description" : "验证手机号是否已被注册",
        "operationId" : "checkPhone",
        "parameters" : [ {
          "name" : "phone",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultBoolean"
                }
              }
            }
          }
        }
      }
    },
    "/user/check/idCard" : {
      "get" : {
        "tags" : [ "用户管理" ],
        "summary" : "检查身份证是否存在",
        "description" : "验证身份证号是否已被注册",
        "operationId" : "checkIdCard",
        "parameters" : [ {
          "name" : "idCard",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultBoolean"
                }
              }
            }
          }
        }
      }
    },
...
    "/contact/list/{userId}" : {
      "get" : {
        "tags" : [ "联系人管理" ],
        "summary" : "获取联系人列表",
        "description" : "获取指定用户的所有常用联系人",
        "operationId" : "getContacts",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultListContact"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "UserInfoDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "username" : {
            "type" : "string"
          },
          "idCard" : {
            "type" : "string"
          },
          "realName" : {
            "type" : "string"
          },
          "phone" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "userType" : {
            "type" : "integer",
            "format" : "int32"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "createTime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updateTime" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "ResultString" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "type" : "string"
          },
          "timestamp" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
      "ResultBoolean" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "type" : "boolean"
          },
          "timestamp" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "WaitlistCreateDTO" : {
        "required" : [ "departureDate", "endStationId", "seatType", "startStationId", "trainId" ],
        "type" : "object",
        "properties" : {
          "trainId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "startStationId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "endStationId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "departureDate" : {
            "type" : "string",
            "format" : "date"
          },
          "seatType" : {
            "type" : "string"
          }
        }
      },
      "ResultWaitlistOrderDTO" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/WaitlistOrderDTO"
          },
          "timestamp" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "WaitlistOrderDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "waitlistNumber" : {
            "type" : "string"
          },
          "userId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "username" : {
            "type" : "string"
          },
          "trainId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "trainNumber" : {
            "type" : "string"
          },
          "startStationId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "startStationName" : {
            "type" : "string"
          },
          "endStationId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "endStationName" : {
            "type" : "string"
          },
          "seatType" : {
            "type" : "string"
          },
          "seatTypeName" : {
            "type" : "string"
          },
          "departureDate" : {
            "type" : "string",
            "format" : "date"
          },
          "status" : {
            "type" : "string"
          },
          "statusText" : {
            "type" : "string"
          },
          "createTime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "expireTime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "orderNumber" : {
            "type" : "string"
          }
        }
      },
      "UserRegisterDTO" : {
        "required" : [ "idCard", "password", "phone", "realName", "username" ],
        "type" : "object",
        "properties" : {
          "username" : {
            "maxLength" : 20,
            "minLength" : 3,
            "type" : "string"
          },
          "password" : {
            "maxLength" : 20,
            "minLength" : 6,
            "type" : "string"
          },
          "idCard" : {
            "pattern" : "\\d{17}[\\d|x|X]|\\d{15}",
            "type" : "string"
          },
          "realName" : {
            "maxLength" : 10,
            "minLength" : 2,
            "type" : "string"
          },
          "phone" : {
            "pattern" : "1[3-9]\\d{9}",
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          }
        }
      },
      "ResultUserInfoDTO" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/UserInfoDTO"
          },
          "timestamp" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "UserLoginDTO" : {
        "required" : [ "password", "username" ],
        "type" : "object",
        "properties" : {
          "username" : {
            "maxLength" : 20,
            "minLength" : 3,
            "type" : "string"
          },
          "password" : {
            "maxLength" : 20,
            "minLength" : 6,
            "type" : "string"
          }
        }
      },
...
      "ResultListWaitlistTicketDTO" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/WaitlistTicketDTO"
            }
          },
          "timestamp" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "WaitlistTicketDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "ticketNumber" : {
            "type" : "string"
          },
          "originalOrderNumber" : {
            "type" : "string"
          },
          "trainId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "trainNumber" : {
            "type" : "string"
          },
          "startStationId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "startStationName" : {
            "type" : "string"
          },
          "endStationId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "endStationName" : {
            "type" : "string"
          },
          "departureDate" : {
            "type" : "string",
            "format" : "date"
          },
          "seatType" : {
            "type" : "string"
          },
          "seatTypeName" : {
            "type" : "string"
          },
          "carriageNumber" : {
            "type" : "string"
          },
          "seatNumber" : {
            "type" : "string"
          },
          "price" : {
            "type" : "number"
          },
          "priceText" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "statusText" : {
            "type" : "string"
          },
          "source" : {
            "type" : "string"
          },
          "createTime" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "Contact" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "userId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "phone" : {
            "type" : "string"
          },
          "idCard" : {
            "type" : "string"
          },
          "type" : {
            "type" : "integer",
            "format" : "int32"
          },
          "isDefault" : {
            "type" : "integer",
            "format" : "int32"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "createTime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updateTime" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "ResultListContact" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Contact"
            }
          },
          "timestamp" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
    }
  }
}